# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Graduation Project Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List users
      parameters:
        - name: role
          in: query
          schema:
            type: string
            enum: [student, supervisor, coordinator]
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /teams:
    get:
      operationId: listTeams
      tags: [teams]
      summary: List all teams
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Teams list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Team"
    post:
      operationId: createTeam
      tags: [teams]
      summary: Create a team
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTeamBody"
      responses:
        "201":
          description: Team created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
  /teams/{id}:
    get:
      operationId: getTeam
      tags: [teams]
      summary: Get team by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Team
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
    put:
      operationId: updateTeam
      tags: [teams]
      summary: Update team
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTeamBody"
      responses:
        "200":
          description: Updated team
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
...
  /teams/{id}/members:
    get:
      operationId: getTeamMembers
      tags: [teams]
      summary: Get team members
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Members list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TeamMember"
  /teams/{id}/leave:
    post:
      operationId: leaveTeam
      tags: [teams]
      summary: Leave a team
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Left team
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List tasks for my team
      responses:
        "200":
          description: Tasks list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a task (supervisor only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskBody"
      responses:
        "201":
          description: Task created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}:
    get:
      operationId: getTask
      tags: [tasks]
      summary: Get task by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    put:
      operationId: updateTask
      tags: [tasks]
      summary: Update task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskBody"
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [student, supervisor, coordinator]
        createdAt:
          type: string
          format: date-time
      required: [id, name, email, role, createdAt]
...
    Team:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        projectTitle:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        status:
          type: string
          enum: [forming, active, supervised, completed]
        leaderId:
          type: integer
        supervisorId:
          type: integer
          nullable: true
        currentPhase:
          type: string
          enum: [proposal, progress, final]
          nullable: true
        createdAt:
          type: string
          format: date-time
        leader:
          $ref: "#/components/schemas/User"
        supervisor:
          $ref: "#/components/schemas/User"
          nullable: true
        memberCount:
          type: integer
      required: [id, name, status, leaderId, createdAt, leader, memberCount]
    CreateTeamBody:
      type: object
      properties:
        name:
          type: string
        projectTitle:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
      required: [name]
    UpdateTeamBody:
      type: object
      properties:
        name:
          type: string
        projectTitle:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
    TeamMember:
      type: object
      properties:
        id:
          type: integer
        teamId:
          type: integer
        userId:
          type: integer
        role:
          type: string
          enum: [leader, member]
        joinedAt:
          type: string
          format: date-time
        user:
          $ref: "#/components/schemas/User"
      required: [id, teamId, userId, role, joinedAt, user]
...
    Task:
      type: object
      properties:
        id:
          type: integer
        teamId:
          type: integer
        title:
          type: string
        description:
          type: string
          nullable: true
        deadline:
          type: string
          format: date-time
          nullable: true
        phase:
          type: string
          enum: [proposal, progress, final]
        status:
          type: string
          enum: [pending, submitted, reviewed]
        createdAt:
          type: string
          format: date-time
        team:
          $ref: "#/components/schemas/Team"
      required: [id, teamId, title, phase, status, createdAt, team]
    CreateTaskBody:
      type: object
      properties:
        teamId:
          type: integer
        title:
          type: string
        description:
          type: string
          nullable: true
        deadline:
          type: string
          format: date-time
          nullable: true
        phase:
          type: string
          enum: [proposal, progress, final]
      required: [teamId, title, phase]
    UpdateTaskBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
          nullable: true
        deadline:
          type: string
          format: date-time
          nullable: true
        status:
          type: string
          enum: [pending, submitted, reviewed]
...